requirements.txt
paho-mqtt>=1.3
flask
pyyaml>=3
